V8 (#3)
* Provide a tool for building V8 heap snapshots from .js files

V8 requires an initial heap image. Instead of providing it as a file
dependency, instead allow it to be linked directly to an executable.
This simplifies deployment. It's also possible to dump the V8 heap that results
from executing JavaScript. Loading this heap image instead of re-parsing the
script can speed up startup time significantly.

This change adds a tool, v8_heap_gen, that can produce libraries with
prebaked V8 heap images.