Compile the TypeScript compiler to a heap image (#4)
* Compile the TypeScript compiler to a heap image

V8 takes a nontrivial amount of time to chew on tsc.js, which we'd have
to suffer every time we run a binary that needs to parse JavaScript.
This PR does this work ahead of time and generates a heap image that
can be quickly loaded into a fresh isolate. There's a single function
exposed ("parse") that will produce a JSON AST from a given input string.
This interface will be expanded in the future.

tsconfig.json.in is compatible with the (eventual) cmake build.