Add code generation for patterns to tt (#26)
This PR also fixes up an issue with the tt
preprocessor wrt BCPL comments. I'd swapped out
the default Bison location class with the
Anodyne one (for consistency). Because Anodyne's
is offset and not line/column based, we do care
how many characters are left on a BCPL comment line.