Quick and dirty port of the OCaml pretty-print module (and only that) (#36)
* Quick and dirty port of the OCaml pretty-print module

This implements a pretty-printing algorithm for dumping
expression trees. It's been translated more or less directly
from the corresponding F#/OCaml code, which is why it lives
in third_party. Some documentation has been reverse-engineered.

We may want to replace this in the future, but it's enough to
produce somewhat readable dumps.